Big refactor coming...
I need to introduce another nest of menus to support 3 level of menus with the final layer only supporting callbacks. I cannot see a way to make this into a dynamic thing as of now, so I will add this to support what I need for now and revisit this in a future zombie issue.. todo make zombie issue for this!